+++ mythtv/libs/libmythtv/recorders/satipchannel.cpp
@@ -49,7 +49,7 @@ bool SatIPChannel::Open(void)
     m_tunerType = SatIP::toTunerType(m_device);
 
     LOG(VB_CHANNEL, LOG_DEBUG, LOC + QString("Open(%1) m_tunerType:%2 %3")
-        .arg(m_device).arg(m_tunerType).arg(m_tunerType.toString()));
+        .arg(m_device).arg(m_tunerType.toInt()).arg(m_tunerType.toString()));
 
     if (!InitializeInput())
     {
